(maint) Rename the module internals
Prior to this commit, the module was named http_authorization.
During the public RFC process, several commenters noted that this
name is overly generic and doesn't accurately reflect the scope
of the module.

This commit renames all the internals to "puppet_authorization",
which was the name suggested by Nick Fagerlund the Nomenclator,
aka "Them What Names Things", and +1'ed by several others.

The spec tests pass as-is but in addition to merging this commit,
the repository itself will need a rename.