(MAINT) Fix set_value to only consider rule name
This dumbs down rule equality checking in the provider's set_value
method to only consider rule['name'] when preventing duplicates.

Prior to this commit, rules would be duplicated in the auth.conf file
because a full object equality comparison was failing to recognize the
value_to_set was already present in the configuration file.

It's possible that a similar change will need to be made to remove_value
and possibly exists?, but those changes are still being decided.