(FM-3807) Add rules when target array does not already exist
This commit allows the provider to create a new array and add a rule to
it even when the rule setting in the target file does not already
contain an array.  In previous commits, an attempt to add a rule in this
scenario would generate an error and no changes would be applied to the
target file.